layout

+++ wp-content/themes/Woodd/category.php
@@ -11,30 +11,21 @@
 
 get_header(); ?>
 
-<?php
-$term_id = 4;
-$taxonomy_name = 'category';
-$termchildren = get_term_children( $term_id, $taxonomy_name );
-
-echo '<ul>';
-foreach ( $termchildren as $child ) {
-	$term = get_term_by( 'id', $child, $taxonomy_name );
-	echo '<li><a href="' . get_term_link( $child, $taxonomy_name ) . '">' . $term->name . '</a></li>';
-}
-echo '</ul>';
-?>
-
-				<?php
-	$terms = get_terms( 'category' , 'parent='.get_query_var('cat') );
+<?php if (function_exists('breadcrumbs')) breadcrumbs(); ?>
 
 
-	 	$terms = get_terms("category");
-	 	$count = count($terms); ?>
+		<?php
+		$queried_object = get_queried_object();
+		$term_id = $queried_object->term_id;
+		$taxonomy_name = 'category';
+		$termchildren = get_term_children( $term_id, $taxonomy_name );
+	 	$count = count($termchildren); ?>
 	 	<ul id="filters">
 	 	<a href="#" data-filter="*">Show all |</a>
 	 	<?php
 	 	if ( $count > 0 ){
-	    foreach ( $terms as $term ) {
+	    foreach ( $termchildren as $child ) {
+	    $term = get_term_by( 'id', $child, $taxonomy_name );
 	    $category_name = $term->name;
 	    $category = $term->slug;
 	    echo '<a href="#" data-filter=".'. $category .'">'. $category_name .' | </a>';
